Remove redundant legacy UI code
