+++ docs/tests.md
@@ -123,7 +123,8 @@ Fast Node tests:
 - `launch_args_unit.cjs`: launch argument generation for inert defaults, default/custom/unavailable model roots, traversal rejection, sampler-related flag behavior, server-wide reasoning-effort template kwargs, model-source recognition, and sensitive-value redaction.
 - `chat_template_selection_unit.cjs`: the extracted chat-template selection facade against the real flag state — load/configure inertness (no flag writes or DOM initialization, including on the detached Chat page), auto/builtin/bundled/custom/unsupported selections, legacy built-in round trips (such as `phi4`) outside the curated dropdown, atomic one-patch mode transitions with exact delete keys and exactly one `afterPatch`/`postUpdate` notification pair per action, including a regression case that rejects duplicate writes, read-only reverse mapping in mixed state (both template fields set, bundled-path precedence), Windows-path normalization for comparison only (no case folding, basename matching, or stored-path rewriting), the manual custom-path clearing rule through the `beforePathPatch` hook (typed, picked, and cleared paths, one write/broadcast each), lookups after `applyFlagValues()` replaces the state object, unsupported stored values keeping their launch warning/omission, and at most one of `--chat-template`/`--chat-template-file` emitted.
 - `output_cursor_unit.cjs`: generation-aware process output cursor consumption, stale-response rejection, and `invalidate()` semantics that preserve the cursor while rejecting in-flight responses.
-- `monitor_ui_unit.cjs`: hermetic Monitor tests for polling and badge stability, card visibility/reordering and focus preservation, inference baselines and telemetry normalization, and safe rendering of hostile telemetry text. Average-speed coverage checks processing-time weighting, idle periods, restored targets, token/time counter rollback, missing timings, and resets before or between valid samples. Live prompt/generation coverage checks updates before request completion, parallel slots, task changes, stalls, missing samples, long polling gaps, and returning to the completed-session average. Prompt cases also verify cache exclusion, null counter handling, excluding intervals that span prefill and generation, and batch updates spanning several polls without spikes or zero-rate flicker.
+- `monitor_ui_unit.cjs`: hermetic Monitor tests for polling and badge stability, card visibility/reordering and focus preservation, shared inference snapshot rendering and partial-source guidance, and safe rendering of hostile telemetry text. Loads the inference core and Monitor in canonical script order.
+- `inference_stats_unit.cjs`: metrics parsing, slot normalization, and inference baselines in a VM without DOM, storage, fetch, or timers; checks inert loading/creation and independent engine instances. Average-speed coverage checks processing-time weighting, idle periods, restored targets, token/time counter rollback, missing timings, and resets before or between valid samples. Live prompt/generation coverage checks updates before request completion, parallel slots, task changes, stalls, missing samples, long polling gaps, and returning to the completed-session average. Prompt cases also verify cache exclusion, null counter handling, excluding intervals that span prefill and generation, and batch updates spanning several polls without spikes or zero-rate flicker.
 - `process_lifecycle_unit.cjs`: guarded launch/stop/switch ordering, readiness progression, generation conflicts, out-of-band replacement reconciliation, refused-stop recovery, stop-during-load, and stale transition handling.
 - `model_switch_ui_unit.cjs`: two-slot persistence, assignment validation, recoverable slot states, cancellation/failure cleanup, active-runtime display precedence, sidebar slider availability/drag thresholds/markup, safe rendering helpers, and storage fallback. Browser interactions cover assignment changes, refresh, and drag/keyboard guards; exact CSS and source-text locks have been removed.
 - `benchmark_args_unit.cjs`: benchmark/perplexity argument adaptation through the shared local-model path builder without mutating source presets, plus visible model-folder load failures in the manual-model selector.
@@ -161,6 +162,8 @@ Fast Node tests:
 Browser smoke test:
 
 Benchmark action coverage checks normal completion and parsed throughput summaries, safe output rendering, launch failure and retry, refused Stop with resumed polling and a successful retry, reconnecting to an existing benchmark, and WikiText preparation failure/retry before perplexity launch. It exercises the real app/lifecycle/benchmark wiring through API fixtures.
+
+Detached Chat integration also verifies that a hidden host continues producing inference snapshots and owns every metrics/slots request while the popup has no inference engine. System telemetry stays paused, and closing the popup stops hidden-host inference polling and clears its timers/controller.
 
 Monitor runtime coverage checks authoritative identity despite pending edits, safe long model names, focused Configure change review, external-server navigation, retained versus empty logs, and layout containment at 900/390px. A missing-vendor-probe fixture verifies useful system readings and an optional, keyboard-operable GPU setup disclosure whose focus/open state survive polling. A delayed-response fixture ignores AbortSignal to verify that epoch invalidation still rejects metrics/slots after reconnecting to the same external endpoint. `monitor_ui_unit.cjs` also covers partial inference guidance, model-only changes, failed actions with a still-active process, and vendor-independent readings.
