+++ ui/index.html
@@ -1901,6 +1901,7 @@
 <script src="/js/chat/chat-main.js?v=chat-split-1"></script>
 <script src="/js/chat-window.js?v=chat-window-1"></script>
 <script src="/js/benchmark-ui.js?v=revamp-2"></script>
+<script src="/js/inference-stats.js?v=1"></script>
 <script src="/js/monitor-ui.js?v=monitor-1"></script>
 <script src="/js/shell-ui.js?v=shell-1"></script>
 <script src="/js/app.js?v=revamp-4"></script>
